Renamed circularTextbox.ts to canvasCircularTextbox.ts
